requirements.txt
asttokens==3.0.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.2.0
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
exceptiongroup==1.2.2
executing==2.2.0
fonttools==4.47.0
idna==3.6
ipyevents==2.0.2
ipympl==0.9.6
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.2
joblib==1.4.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lazy_loader==0.3
lxml==5.3.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
mne==1.7.0
mpmath==1.3.0
nibabel==5.2.1
nilearn==0.10.4
numpy==1.26.2
opencv-python==4.10.0.84
packaging==23.2
pandas==1.5.2
pandas-flavor==0.6.0
parso==0.8.4
patsy==0.5.6
Pillow==10.1.0
pingouin==0.5.5
platformdirs==4.1.0
pooch==1.8.0
prompt_toolkit==3.0.50
ptitprince==0.2.7
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.1.1
PyQt5==5.15.9
pyqt5-plugins==5.15.9.2.3
PyQt5-Qt5==5.15.2
pyqt5-tools==5.15.9.3.3
PyQt5_sip==12.17.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
pyvista==0.44.2
pyvistaqt==0.11.2
qt5-applications==5.15.2.2.3
qt5-tools==5.15.2.1.3
QtPy==2.4.3
requests==2.31.0
rpy2==3.5.16
scikit-learn==1.5.2
scipy==1.11.4
scooby==0.10.0
seaborn==0.11.0
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.3
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tqdm==4.66.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
vtk==9.3.1
wcwidth==0.2.13
webencodings==0.5.1
widgetsnbextension==4.0.13
xarray==2025.1.1